# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UNZA DigiHealth API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /queue:
    get:
      operationId: getQueue
      tags: [queue]
      summary: Get current queue
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Queue list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueEntry"
  /queue/my-position:
    get:
      operationId: getMyQueuePosition
      tags: [queue]
      summary: Get my queue position
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Queue position info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueuePosition"
  /queue/join:
    post:
      operationId: joinQueue
      tags: [queue]
      summary: Join the queue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinQueueRequest"
      responses:
        "201":
          description: Joined queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
  /queue/{id}/complete:
    patch:
      operationId: completeQueueEntry
      tags: [queue]
      summary: Mark queue entry as complete (doctor)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"

  # Prescriptions
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: List prescriptions
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/PrescriptionStatus"
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"

    post:
      operationId: createPrescription
      tags: [prescriptions]
      summary: Create a prescription (doctor)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrescriptionRequest"
      responses:
        "201":
          description: Created prescription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /prescriptions/{id}:
    get:
      operationId: getPrescription
      tags: [prescriptions]
      summary: Get prescription by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prescription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /prescriptions/{id}/dispense:
    patch:
      operationId: dispensePrescription
      tags: [prescriptions]
      summary: Dispense prescription (pharmacist)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispensePrescriptionRequest"
      responses:
        "200":
          description: Dispensed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"

  # Lab Requests
  /lab-requests:
    get:
      operationId: listLabRequests
      tags: [lab]
      summary: List lab requests
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/LabRequestStatus"
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of lab requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabRequest"

    post:
      operationId: createLabRequest
      tags: [lab]
      summary: Create a lab request (doctor)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabRequestRequest"
      responses:
        "201":
          description: Created lab request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabRequest"
  /lab-requests/{id}:
    get:
      operationId: getLabRequest
      tags: [lab]
      summary: Get lab request by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lab request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabRequest"
  /lab-requests/{id}/status:
    patch:
      operationId: updateLabRequestStatus
      tags: [lab]
      summary: Update lab request status (lab tech)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLabRequestStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabRequest"

  # Lab Results
  /lab-results:
    post:
      operationId: uploadLabResult
      tags: [lab]
      summary: Upload lab result (lab tech)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabResultRequest"
      responses:
        "201":
          description: Uploaded lab result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabResult"
  /lab-results/{requestId}:
    get:
      operationId: getLabResult
      tags: [lab]
      summary: Get lab result for a request
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lab result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabResult"

  # Notifications
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRole:
      type: string
      enum: [student, doctor, pharmacist, lab_technician, mental_health_counselor, hiv_professional, admin]
    UserStatus:
      type: string
      enum: [active, inactive, suspended]
    ConsultationStatus:
      type: string
      enum: [submitted, under_review, assigned, responded, closed]
    PrescriptionStatus:
      type: string
      enum: [pending, dispensed, cancelled]
    LabRequestStatus:
      type: string
      enum: [pending, in_progress, completed, cancelled]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        studentNumber:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, status, createdAt]
...
    Consultation:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        doctorId:
          type: integer
        symptoms:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        attachments:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/ConsultationStatus"
        diagnosis:
          type: string
        notes:
          type: string
        student:
          $ref: "#/components/schemas/User"
        doctor:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, studentId, symptoms, severity, status, createdAt, updatedAt]
...
    QueueEntry:
      type: object
      properties:
        id:
          type: integer
        consultationId:
          type: integer
        studentId:
          type: integer
        queueNumber:
          type: integer
        status:
          type: string
          enum: [waiting, in_progress, completed, skipped]
        estimatedWaitMinutes:
          type: integer
        student:
          $ref: "#/components/schemas/User"
        consultation:
          $ref: "#/components/schemas/Consultation"
        createdAt:
          type: string
          format: date-time
      required: [id, consultationId, studentId, queueNumber, status, createdAt]
    QueuePosition:
      type: object
      properties:
        position:
          type: integer
        queueNumber:
          type: integer
        estimatedWaitMinutes:
          type: integer
        status:
          type: string
          enum: [waiting, in_progress, completed, skipped, not_in_queue]
        totalInQueue:
          type: integer
      required: [position, status, totalInQueue]

    # Prescription
    Prescription:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        consultationId:
          type: integer
        medication:
          type: string
        dosage:
          type: string
        instructions:
          type: string
        duration:
          type: string
        status:
          $ref: "#/components/schemas/PrescriptionStatus"
        dispensedAt:
          type: string
          format: date-time
        dispensedBy:
          type: integer
        patient:
          $ref: "#/components/schemas/User"
        doctor:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
      required: [id, patientId, doctorId, medication, dosage, instructions, status, createdAt]
    CreatePrescriptionRequest:
      type: object
      properties:
        patientId:
          type: integer
        consultationId:
          type: integer
        medication:
          type: string
        dosage:
          type: string
        instructions:
          type: string
        duration:
          type: string
      required: [patientId, medication, dosage, instructions]

    # Lab
    LabRequest:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        consultationId:
          type: integer
        testType:
          type: string
        notes:
          type: string
        status:
          $ref: "#/components/schemas/LabRequestStatus"
        patient:
          $ref: "#/components/schemas/User"
        doctor:
          $ref: "#/components/schemas/User"
        result:
          $ref: "#/components/schemas/LabResult"
        createdAt:
          type: string
          format: date-time
      required: [id, patientId, doctorId, testType, status, createdAt]
    CreateLabRequestRequest:
      type: object
      properties:
        patientId:
          type: integer
        consultationId:
          type: integer
        testType:
          type: string
        notes:
          type: string
      required: [patientId, testType]
    LabResult:
      type: object
      properties:
        id:
          type: integer
        requestId:
          type: integer
        results:
          type: string
        attachment:
          type: string
        uploadedBy:
          type: integer
        uploadedAt:
          type: string
          format: date-time
      required: [id, requestId, results, uploadedAt]
    CreateLabResultRequest:
      type: object
      properties:
        requestId:
          type: integer
        results:
          type: string
        attachment:
          type: string
      required: [requestId, results]

    # Notification
...
    JoinQueueRequest:
      type: object
      properties:
        consultationId:
          type: integer
      required: [consultationId]
    DispensePrescriptionRequest:
      type: object
      properties:
        dispensedBy:
          type: integer
        notes:
          type: string
      required: [dispensedBy]
    UpdateLabRequestStatusRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/LabRequestStatus"
      required: [status]
...