# lib/api-zod/src/generated/types/labResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UNZA DigiHealth API
 * OpenAPI spec version: 0.1.0
 */

export interface LabResult {
  id: number;
  requestId: number;
  results: string;
  attachment?: string;
  uploadedBy?: number;
  uploadedAt: Date;
}